Add @JsonIgnore to getter
Because DeviceRegistration uses @JsonProperty to mark fields as getters
and setters for the JSON serializer, the getter methods should be
ignored. If `#getAttestationCertificate` is not marked as ignorable,
Jackson will attempt to serialize an X509Certificate instance, which
will fail. Instead the string representation of the certificate should
be serialized.

This is not a problem in this library, because `#toJson` is called
directly, but other projects depending on this class to serialize neatly
with Jackson need the extra annotations.