Make attestationTrusted() implementation a little less misleading
The previous filter-isPresent method made it look like it was only checking for
the presence of a "trusted" attribute, and would also accept a present false
value as meaning the attestation is trusted. This is in fact not a bug, but the
map-orElse pattern makes it clearer what is actually being checked.