Remove defaults for userVerification and residentKey
By having defaults in the library, we are inadvertently suppressing warnings
that browsers may issue in the browser console when for example
`userVerification` is not set explicitly. By making the defaults in the library
not set, we fall back to the browser defaults and do not suppress any such
warnings.