Drop U2F registration from demo app
Chrome has fully dropped support for the U2F API, so this is unlikely
to work anymore.

The main purpose of this feature in the demo was to illustrate use of
the `appid` and `appidExclude` extensions.

See: https://groups.google.com/a/chromium.org/g/blink-dev/c/xHC3AtU_65A