Set setPolicyQualifiersRejected to false for testing
The default Java certificate path validator rejects certificates with critical
policy extensions, and Windows Hello uses such an attestation cert. The solution
for this is to set the `setPolicyQualifiersRejected(boolean)` setting and for
the application to validate the policy tree. For now, we'll just set the
parameter to `false` and add a validator setting later.