Go back to JDK 11 for mutation tests
When running the mutation tests in JDK 17, they seem to incorrectly report lower
coverage than what is actually the case - including whole classes plummeting to
0% coverage. I have manually confirmed that some of the mutations reported as
uncovered are in fact covered by the tests. Perhaps there's some test discovery
mechanism that breaks in JDK 17; I'll have to test newer versions of the test
frameworks and see if that helps.